Refactor: Move RulesScreen to LobbyScreens and implement Single Activity Architecture
