+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <application
@@ -14,14 +13,8 @@
         android:theme="@style/Theme.DoomHamsters"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".RulesActivity"
-            android:exported="false"
-            android:label="@string/title_activity_rules"
-            android:theme="@style/Theme.DoomHamsters" />
-        <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.DoomHamsters">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
