--- app/src/main/java/com/doomhamsters/RulesActivity.kt
@@ -1,119 +1,0 @@
-package com.doomhamsters
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.doomhamsters.ui.theme.DarkBrown
-import com.doomhamsters.ui.theme.DoomHamstersTheme
-import com.doomhamsters.ui.theme.Orange
-import com.doomhamsters.ui.theme.SoftWhite
-import com.doomhamsters.ui.theme.WarmAlmond
-
-class RulesActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            DoomHamstersTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    RulesScreen(
-                        modifier = Modifier.padding(innerPadding),
-                        onBackClick = { finish() }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun RulesScreen(modifier: Modifier = Modifier, onBackClick: () -> Unit) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(WarmAlmond)
-            .padding(24.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = stringResource(R.string.rules_title),
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            color = Orange
-        )
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        Surface(
-            shape = RoundedCornerShape(16.dp),
-            color = Color.White.copy(alpha = 0.5f), // Dein 80% Weiß
-            modifier = Modifier
-                .fillMaxWidth()
-                .shadow(elevation = 1.dp, shape = RoundedCornerShape(16.dp))
-        ) {
-            Column(
-                modifier = Modifier.padding(16.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.rules_content),
-                    fontSize = 18.sp,
-                    //fontWeight = FontWeight.Bold,
-                    color = DarkBrown
-                )
-            }
-        }
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        Button(
-            onClick = onBackClick,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp),
-            shape = RoundedCornerShape(50),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Orange,
-                contentColor = Color.White
-            ),
-            elevation = ButtonDefaults.buttonElevation(defaultElevation = 3.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.rules_back),
-                fontSize = 18.sp,
-                fontWeight = FontWeight.Bold,
-                color = SoftWhite
-            )
-        }
-    }
-}
-@Preview(showBackground = true)
-@Composable
-fun RulesPreview() {
-    DoomHamstersTheme {
-        RulesScreen(onBackClick = {})
-    }
-}
